Refactor Brs_021
