+++ source/ProcessManager.Orchestrations/Processes/BRS_021/ForwardMeteredData/V1/Orchestration_Brs_021_ForwardMeteredData_V1.cs
@@ -13,8 +13,10 @@
 // limitations under the License.
 
 using Energinet.DataHub.ElectricityMarket.Integration;
+using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationDescription;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Core.Infrastructure.Extensions.DurableTask;
+using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
 using Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ForwardMeteredData.V1.Activities;
 using Microsoft.Azure.Functions.Worker;
@@ -29,7 +31,7 @@ internal class Orchestration_Brs_021_ForwardMeteredData_V1
     internal const int FindReceiverStep = 3;
     internal const int EnqueueActorMessagesStep = 4;
 
-    public static readonly Brs_021_ForwardedMeteredData_V1 UniqueName = new();
+    public static readonly OrchestrationDescriptionUniqueNameDto UniqueName = Brs_021_ForwardedMeteredData.V1;
 
     private readonly TaskOptions _defaultRetryOptions;
